remove icon components for svg strings
