--- src/ui/components/icons/Expand.svelte
@@ -1,15 +1,0 @@
-<svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    stroke-width="2"
-    stroke-linecap="round"
-    stroke-linejoin="round"
-    ><polyline points="15 3 21 3 21 9" /><polyline points="9 21 3 21 3 15" /><line
-        x1="21"
-        y1="3"
-        x2="14"
-        y2="10"
-    /><line x1="3" y1="21" x2="10" y2="14" /></svg
->
