--- src/ui/components/icons/FileCode.svelte
@@ -1,12 +1,0 @@
-<svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    stroke-width="2"
-    stroke-linecap="round"
-    stroke-linejoin="round"
-    ><path d="M4 22h14a2 2 0 0 0 2-2V7.5L14.5 2H6a2 2 0 0 0-2 2v4" /><polyline
-        points="14 2 14 8 20 8"
-    /><path d="m9 18 3-3-3-3" /><path d="m5 12-3 3 3 3" /></svg
->
